VEC-69: (Frontend) reordered components
