+++ vectre/frontend/src/components/Modals/BannerProfileEditPicsWrapper/ProfilePic/ProfilePic.js
@@ -1,22 +1,16 @@
 import {
-    Flex,
     Box,
     Image,
-    Spacer,
     IconButton
 } from "@chakra-ui/react"
 import { BsFillPencilFill } from "react-icons/bs";
-import { getAvatarOrDefault, getBannerOrDefault } from "../../../utils/Utils";
+import { getAvatarOrDefault } from "../../../../utils/Utils";
 
-const BannerProfileEditPicsWrapper = ({
+const ProfilePic = ({
     data
 }) => {
     return (
-        <Flex
-            flexDirection={'column'}
-            gap={'20px'}
-            position={'relative'}
-            zIndex={5}>
+        <>
             <Box
                 position={'absolute'}
                 top={'53%'}
@@ -40,28 +34,8 @@ const BannerProfileEditPicsWrapper = ({
                     icon={<BsFillPencilFill size={'1.2rem'} />}
                 />
             </Box>
-            <Image
-                src={getBannerOrDefault(data.banner)}
-                fit={'cover'}
-                overflow={'hidden'}
-                borderRadius={'6px'}
-                height={'200px'} />
-            <IconButton
-                _focus={{ outline: 0 }}
-                bg={'rgba(35, 52, 71, 0.85)'}
-                color={'white'}
-                borderRadius={'40px'}
-                width={'120px'}
-                bottom={'10%'}
-                right={'0%'}
-                mr={'8px'}
-                mb={'5px'}
-                position={'absolute'}
-                icon={<BsFillPencilFill size={'1.5rem'} />}
-            />
-            <Spacer />
-        </Flex>
+        </>
     )
 }
 
-export default BannerProfileEditPicsWrapper;
+export default ProfilePic;
